Compile for windows and linking succeedes.
Everything has been fixed so that the program links. [Installing
Win32Ada was required, dl from libre.adacore.com]

The program now issues an error on run:
This application has requested the Runtime to terminate it in an unusual
way.
Please contact the application's support team for more information.

Oddly, in between actually having the current running program and the
last program I did get it to link and run. There were issues with
nonexistant files & directories but generating empty documents for them
did enable the program to log an attempt to connect to the network.